Refactoring AI Review stuff into a common component area
