+++ src/components/AIReview/index.ts
@@ -15,10 +15,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from "./Game";
-export * from "./GameChat";
-export * from "./GameHooks";
-export * from "./GameTimings";
-export * from "./goban_context";
-export * from "./util";
-export * from "./GameLog";
+export * from "./AIDemoReview";
+export * from "./AISummaryTable";
+export * from "./AIReview";
+export * from "./AIReviewChart";
+export * from "./MoveListPopover";
